requirements.txt
pandas>=1.5.0
psycopg2-binary>=2.9.5
SQLAlchemy>=1.4.0
python-dotenv>=0.19.0
requests>=2.26.0
tqdm>=4.62.0
PyYAML>=6.0
python-dateutil>=2.8.2
pytz>=2020.1
tzdata>=2022.7
typing-extensions>=4.6.0
charset-normalizer>=2.0.0
idna>=2.5
urllib3>=1.21.1
certifi>=2017.4.17